+++ bin/teq.mjs
@@ -4,9 +4,9 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import process from 'node:process';
-import Container from '@teqfw/di/src/Container.mjs';
-import NamespaceRegistry from '@teqfw/di/src/Config/NamespaceRegistry.mjs';
-import PackageGraph from '../src/Infra/PackageGraph.mjs';
+import Container from '@teqfw/di';
+import NamespaceRegistry from '@teqfw/di/node/registry/namespace';
+import PackageRegistry from '@teqfw/di/node/registry/package';
 import ProviderRegistry from '../src/Registry/Provider.mjs';
 import Io from '../src/Adapter/Io.mjs';
 import Bootstrap from '../src/Bootstrap.mjs';
@@ -14,7 +14,7 @@ import Bootstrap from '../src/Bootstrap.mjs';
 const appRoot = path.resolve(process.cwd());
 const bootstrap = new Bootstrap({
     namespaceRegistry: new NamespaceRegistry({fs, path, appRoot}),
-    providerRegistry: new ProviderRegistry({packageGraph: new PackageGraph({fs, path, appRoot})}),
+    providerRegistry: new ProviderRegistry({packageRegistry: new PackageRegistry({fs, path, appRoot})}),
     container: new Container(),
     io: new Io({processModule: {default: process}}),
 });